Revert "commenting segment type perms until FE changes begin in >5.14. Added migration to restore old ctl perms and remove new segment type perms."
This reverts commit 173064d6de1fbfdc7bfd33f8285e6053c81ec725.

+++ segment/api/views/custom_segment/segment_delete.py
@@ -7,6 +7,8 @@ from segment.models.constants import SegmentActionEnum
 from segment.models.constants import SegmentTypeEnum
 from segment.models.utils.segment_action import segment_action
 from segment.utils.utils import AdminCustomSegmentOwnerPermission
+from segment.utils.utils import CustomSegmentChannelDeletePermission
+from segment.utils.utils import CustomSegmentVideoDeletePermission
 from utils.permissions import or_permission_classes
 
 from segment.api.mixins import SegmentTypePermissionMixin
@@ -18,23 +20,19 @@ class SegmentDeleteApiView(DestroyAPIView, SegmentTypePermissionMixin):
     permission_classes = (
          or_permission_classes(
              AdminCustomSegmentOwnerPermission,
-             StaticPermissions.has_perms(StaticPermissions.BUILD__CTL_DELETE),
-             # TODO re-enable after 5.14 for FE work
-             # StaticPermissions.has_perms(StaticPermissions.BUILD__CTL_DELETE_CHANNEL_LIST),
-             # StaticPermissions.has_perms(StaticPermissions.BUILD__CTL_DELETE_VIDEO_LIST),
+             StaticPermissions.has_perms(StaticPermissions.BUILD__CTL_DELETE_CHANNEL_LIST),
+             StaticPermissions.has_perms(StaticPermissions.BUILD__CTL_DELETE_VIDEO_LIST),
          ),
     )
-    # TODO re-enable after 5.14 for FE work
-    # permission_by_segment_type = {
-    #     SegmentTypeEnum.VIDEO.value: StaticPermissions.BUILD__CTL_DELETE_VIDEO_LIST,
-    #     SegmentTypeEnum.CHANNEL.value: StaticPermissions.BUILD__CTL_DELETE_CHANNEL_LIST
-    # }
+    permission_by_segment_type = {
+        SegmentTypeEnum.VIDEO.value: StaticPermissions.BUILD__CTL_DELETE_VIDEO_LIST,
+        SegmentTypeEnum.CHANNEL.value: StaticPermissions.BUILD__CTL_DELETE_CHANNEL_LIST
+    }
 
     @segment_action(SegmentActionEnum.DELETE.value)
     def delete(self, request, *args, **kwargs):
         segment = self.get_object()
-        # TODO re-enable after 5.14 for FE work
-        # self.check_segment_type_permissions(request=request, segment_type=segment.segment_type)
+        self.check_segment_type_permissions(request=request, segment_type=segment.segment_type)
         if segment.audit_id:
             raise ValidationError("Vetted lists can not be deleted.")
         segment.s3.delete_export()
